utils: Fix compiler warnings for comparison of different signedness
Fixes 172 warnings of:
- warning: comparison of integer expressions of different signedness:
  ‘int’ and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]
- warning: comparison of integer expressions of different signedness:
  ‘int’ and ‘unsigned int’ [-Wsign-compare]

+++ utils/sctp/testlib/sctputil.c
@@ -97,7 +97,7 @@ test_print_message(int sk, struct msghdr *msg, size_t msg_len)
 {
 	sctp_cmsg_data_t *data;
 	struct cmsghdr *cmsg;
-	int i;
+	unsigned int i;
 	int done = 0;
 	char save;
 	union sctp_notification *sn;
@@ -117,7 +117,7 @@ test_print_message(int sk, struct msghdr *msg, size_t msg_len)
 		printf("DATA(%d):  ", msg_len);
 		while ( msg_len > 0 ) {
 			char *text;
-			int len;
+			unsigned int len;
 
 			text = msg->msg_iov[index].iov_base;
 			len = msg->msg_iov[index].iov_len;
@@ -248,7 +248,7 @@ test_check_buf_data(void *buf, int datalen, int msg_flags,
 			 "Got a datamsg of unexpected length:%d, expected length:%d",
 			  datalen, expected_datalen);
 
-	if ((msg_flags & ~0x80000000) != expected_msg_flags)
+	if (((int)(msg_flags & ~0x80000000)) != expected_msg_flags)
 		tst_brkm(TBROK, tst_exit,
 			 "Unexpected msg_flags:0x%x expecting:0x%x",
 			  msg_flags, expected_msg_flags);
